Move Docker bridge heuristic functions to a separate file
