+++ internal/executor/rpc/rpc.go
@@ -9,8 +9,6 @@ import (
 	"github.com/cirruslabs/cirrus-cli/internal/executor/build/commandstatus"
 	"github.com/cirruslabs/echelon"
 	"github.com/cirruslabs/echelon/renderers"
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/client"
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/google/uuid"
 	"google.golang.org/grpc"
@@ -90,56 +88,6 @@ func (r *RPC) ServerSecret() string {
 
 func (r *RPC) ClientSecret() string {
 	return r.clientSecret
-}
-
-func getDockerBridgeInterface(ctx context.Context) string {
-	const assumedBridgeInterface = "docker0"
-
-	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		return assumedBridgeInterface
-	}
-	defer cli.Close()
-
-	network, err := cli.NetworkInspect(ctx, "bridge", types.NetworkInspectOptions{})
-	if err != nil {
-		return assumedBridgeInterface
-	}
-
-	bridgeInterface, ok := network.Options["com.docker.network.bridge.name"]
-	if !ok {
-		return assumedBridgeInterface
-	}
-
-	return bridgeInterface
-}
-
-func getDockerBridgeIP(ctx context.Context) string {
-	// Worst-case scenario, but still better than nothing
-	// since there's still a chance this would work with
-	// a Docker daemon configured by default.
-	const assumedBridgeIP = "172.17.0.1"
-
-	iface, err := net.InterfaceByName(getDockerBridgeInterface(ctx))
-	if err != nil {
-		return assumedBridgeIP
-	}
-
-	addrs, err := iface.Addrs()
-	if err != nil {
-		return assumedBridgeIP
-	}
-
-	if len(addrs) != 0 {
-		ip, _, err := net.ParseCIDR(addrs[0].String())
-		if err != nil {
-			return assumedBridgeIP
-		}
-
-		return ip.String()
-	}
-
-	return assumedBridgeIP
 }
 
 // Start creates the listener and starts RPC server in a separate goroutine.
